Batalhalhasdecartas.Nívelmestre_Tema1.c

+++ CartasSuperTrunfo.c
@@ -7,14 +7,13 @@ int main() {
     char Estado1[50], Estado2[50]; 
     char Codigo1[20], Codigo2[20];
     char Cidade1[50], Cidade2[50];
-    int Populacao1, Populacao2;
+    unsigned long int Populacao1, Populacao2;  
     float Area1, Area2;
     float PIB1, PIB2;
     int PontosTuristicos1, PontosTuristicos2;
     float PIBperCapita1, PIBperCapita2;
     float DensidadePopulacional1, DensidadePopulacional2; 
     float SuperPoder1, SuperPoder2;
-
 
     // Dados para a Carta 1
     printf("\nCarta 1:\n");
@@ -28,35 +27,33 @@ int main() {
     scanf("%s", Cidade1);
 
     printf("Digite o número da População: \n");
-    scanf("%d", &Populacao1);
+    scanf("%lu", &Populacao1);  
 
     printf("Digite a Area (em km²): \n");
     scanf("%f", &Area1);
 
     printf("Digite o PIB (em bilhões): \n");
-    scanf("%f", &PIB1);
+    scanf("%f", &PIB1);  
 
     printf("Digite o número de pontos turísticos: \n");
     scanf("%d", &PontosTuristicos1);
 
     // Calculando PIB per capita e Densidade Populacional da Carta 1
-    PIBperCapita1 = PIB1 * 1e9 / Populacao1;  // Multiplicado PIB por 1 bilhão para voltar para reais
+    PIBperCapita1 = PIB1 * 1e9 / Populacao1;  // Convertemos para reais por pessoa
     DensidadePopulacional1 = Populacao1 / Area1;
 
-// Calculando Super Poder para a Carta 1
-SuperPoder1 = Populacao1 + Area1 + PIB1 + PontosTuristicos1 + PIBperCapita1 + (1 / DensidadePopulacional1);
+    // Calculando Super Poder para a Carta 1
+    SuperPoder1 = Populacao1 + Area1 + PIB1 + PontosTuristicos1 + PIBperCapita1 + (1 / DensidadePopulacional1);
 
-
-    // Dados para Carta 1
+    // Dados para a Carta 1
     printf("\n***Carta 1***\n");
     printf("Estado: %s - Codigo: %s - Cidade: %s\n", Estado1, Codigo1, Cidade1);
-    printf("População: %d\n", Populacao1);
-    printf("Área: %.2f km² - PIB: %.2f bilhões\n", Area1, PIB1);  
+    printf("População: %lu\n", Populacao1);  
+    printf("Área: %.2f km² - PIB: %.2f bilhões\n", Area1, PIB1);  // Exibindo PIB em bilhões
     printf("Pontos Turísticos: %d\n", PontosTuristicos1);
-    printf("PIB per Capita: %.2f\n", PIBperCapita1);
+    printf("PIB per Capita: %.2f\n", PIBperCapita1);  // Exibindo PIB per capita com 2 casas decimais
     printf("Densidade Populacional: %.2f\n", DensidadePopulacional1);
     printf("Super Poder: %.2f\n", SuperPoder1);
-
 
     // Dados para a Carta 2
     printf("\nCarta 2:\n");
@@ -70,7 +67,7 @@ SuperPoder1 = Populacao1 + Area1 + PIB1 + PontosTuristicos1 + PIBperCapita1 + (1
     scanf("%s", Cidade2);
 
     printf("Digite o número da População: \n");
-    scanf("%d", &Populacao2);
+    scanf("%lu", &Populacao2);  
 
     printf("Digite a Área (em km²): \n");
     scanf("%f", &Area2);
@@ -82,20 +79,19 @@ SuperPoder1 = Populacao1 + Area1 + PIB1 + PontosTuristicos1 + PIBperCapita1 + (1
     scanf("%d", &PontosTuristicos2);
 
     // Calculando PIB per capita e Densidade Populacional para Carta 2
-    PIBperCapita2 = PIB2 * 1e9 / Populacao2;  // Multiplicando PIB por 1 bilhão para voltar para reais
+    PIBperCapita2 = PIB2 * 1e9 / Populacao2;  
     DensidadePopulacional2 = Populacao2 / Area2;
 
     // Calculando Super Poder para a Carta 2
     SuperPoder2 = Populacao2 + Area2 + PIB2 + PontosTuristicos2 + PIBperCapita2 + (1 / DensidadePopulacional2);
 
-
-    // Dados para Carta 2
+    // Dados para a Carta 2
     printf("\n***Carta 2***\n");
     printf("Estado: %s - Codigo: %s - Cidade: %s\n", Estado2, Codigo2, Cidade2);
-    printf("População: %d\n", Populacao2);
-    printf("Área: %.2f km² - PIB: %.2f bilhões\n", Area2, PIB2);  
+    printf("População: %lu\n", Populacao2);  
+    printf("Área: %.2f km² - PIB: %.2f bilhões\n", Area2, PIB2);  // Exibindo PIB em bilhões
     printf("Pontos Turísticos: %d\n", PontosTuristicos2);
-    printf("PIB per Capita: %.2f\n", PIBperCapita2);
+    printf("PIB per Capita: %.2f\n", PIBperCapita2);  // Exibindo PIB per capita com 2 casas decimais
     printf("Densidade Populacional: %.2f\n", DensidadePopulacional2);
     printf("Super Poder: %.2f\n", SuperPoder2);
 
@@ -122,7 +118,6 @@ SuperPoder1 = Populacao1 + Area1 + PIB1 + PontosTuristicos1 + PIBperCapita1 + (1
 
     // Comparação do Super Poder
     printf("Super Poder: %.2f > %.2f? %d\n", SuperPoder1, SuperPoder2, SuperPoder1 > SuperPoder2);
-
 
     return 0;
 }
